requirements.txt
openai
python-dotenv
pyperclip
numpy
pyaudio
lameenc
keyboard
soundfile
pyaudio
